+++ IsoSpec++/marginalTrek++.h
@@ -501,7 +501,7 @@ class SingleAtomMarginal : public Marginal
     const pod_vector<double>& conf_lprobs() const { return lProbs; }
     const pod_vector<double>& conf_masses() const { return masses; }
 
-    int get_original_position(int idx) const
+    size_t get_original_position(size_t idx) const
     {
         #ifdef ISOSPEC_DEBUG
         if (idx < 0 || static_cast<size_t>(idx) >= original_indexes.size())
